feat: BeerPurchase component added
