+++ common/ReviewModal/ReviewModal.tsx
@@ -2,16 +2,11 @@ import {
   Box,
   Button,
   Flex,
-  HStack,
-  IconButton,
-  Input,
   Modal,
   ModalBody,
   ModalContent,
-  ModalContentProps,
   ModalFooter,
   ModalHeader,
-  Tag,
   Text,
   Textarea,
   useDisclosure,
@@ -19,20 +14,13 @@ import {
 } from "@chakra-ui/react";
 import React, {ChangeEvent, useState} from "react";
 import {ReviewStatic} from "../../interface/static";
-import {BeerResultType, CategoryType, ReviewType} from "../../interface/types";
-import {
-  CrossXBlack,
-  EditPencil,
-  OrangeCamera,
-  RightArrow,
-  WhiteCross,
-} from "../../public/svg";
-import SearchInput from "../../src/search/SearchInput";
+import {ReviewType} from "../../interface/types";
+import {EditPencil, OrangeCamera} from "../../public/svg";
 import BottomDrawer from "../BottomDrawer";
-import {LeftBackRandom} from "../headers/LeftBackRandom";
 import {LeftCloseRandom} from "../headers/LeftCloseRandom";
 import {Rating} from "../Rating";
 import BeerName from "./BeerName";
+import {BeerPurchase} from "./BeerPurchase";
 import {BeerSearchContent} from "./BeerSearchContent";
 
 export const ReviewModal = () => {
@@ -95,7 +83,6 @@ export const ReviewModal = () => {
         {/* TODO: should be replaced */}
         <EditPencil />
       </Button>
-      {/* LOGOUT drawer */}
       <BottomDrawer
         headerText={"정말로 나가실 건가요?"}
         onClose={CloseReviewDrawer.onClose}
@@ -136,6 +123,7 @@ export const ReviewModal = () => {
           },
         }}
       />
+
       <Modal onClose={onClose} size={"full"} isOpen={isOpen}>
         {step === 0 && (
           <ModalContent px="20px" pb="40px" maxW="452px">
@@ -165,103 +153,16 @@ export const ReviewModal = () => {
                     rate={reviewInfo.rate}
                   />
                 </Flex>
+
                 {/* purchase */}
-                <VStack p="10px" gap="10px" w="full" alignItems={"flex-start"}>
-                  <Box>
-                    <Text as="span" textStyle="h2" textColor="black.100">
-                      이 맥주 어디서 구매하셨나요?{" "}
-                    </Text>
-                    <Text as="span" textStyle="h2" textColor="gray.200">
-                      (선택)
-                    </Text>
-                  </Box>
-                  <HStack w="full" gap="10px" justify={"space-between"}>
-                    {reviewInfo.place ? (
-                      <HStack>
-                        <Tag
-                          justifyContent="row"
-                          borderRadius={"4px"}
-                          bg="white"
-                          border="1px solid"
-                          borderColor={"orange.200"}
-                          cursor="pointer"
-                        >
-                          <Text textStyle={"h2"} textColor="orange.200">
-                            {reviewInfo.place}
-                          </Text>
-                        </Tag>
-                        {reviewInfo.place === "기타" && (
-                          <HStack
-                            px={"0px"}
-                            py={"2px"}
-                            borderBottom="1px solid"
-                            borderBottomColor={"orange.200"}
-                          >
-                            {/* styling slightly weird, width should be fixed */}
-                            <Input
-                              onChange={handleChangePlace}
-                              value={placeInputValue}
-                              w="auto"
-                              h="auto"
-                              border="none"
-                              px={"0px"}
-                              borderRadius={"none"}
-                              focusBorderColor="none"
-                              placeholder="직접 입력해주세요!"
-                              _placeholder={{color: "orange.100"}}
-                              color="orange.200"
-                              textStyle={"h2"}
-                            />
+                <BeerPurchase
+                  reviewInfo={reviewInfo}
+                  handleChangePlace={handleChangePlace}
+                  clearInput={clearInput}
+                  handleClickPlaceTag={handleClickPlaceTag}
+                  placeInputValue={placeInputValue}
+                />
 
-                            <IconButton
-                              onClick={clearInput}
-                              bg="gray.200"
-                              size={"24px"}
-                              borderRadius="full"
-                              px={"0px"}
-                              aria-label="delete-x-button"
-                              icon={<WhiteCross />}
-                              _hover={{}}
-                              _active={{}}
-                            />
-                          </HStack>
-                        )}
-                      </HStack>
-                    ) : (
-                      <Flex gap="10px">
-                        {purchasePlaces.map((place) => {
-                          return (
-                            <Tag
-                              borderRadius={"4px"}
-                              bg="gray.100"
-                              key={place}
-                              size="md"
-                              cursor="pointer"
-                              onClick={() => handleClickPlaceTag(place)}
-                            >
-                              <Text textStyle={"h2"} textColor="gray.300">
-                                {place}
-                              </Text>
-                            </Tag>
-                          );
-                        })}
-                      </Flex>
-                    )}
-
-                    {reviewInfo.place && (
-                      <IconButton
-                        bg="initial"
-                        size={"24px"}
-                        px={"0px"}
-                        aria-label="delete-x-button"
-                        icon={<CrossXBlack />}
-                        _hover={{}}
-                        _active={{}}
-                        onClick={() => handleClickPlaceTag(null)}
-                      />
-                    )}
-                  </HStack>
-                </VStack>
                 {/* review */}
                 <VStack p="10px" w="full" alignItems={"flex-start"} gap="10px">
                   <Box>
@@ -366,6 +267,3 @@ export const ReviewModal = () => {
     </Box>
   );
 };
-
-//TODO : should be replaced
-export const purchasePlaces = ["편의점", "펍", "대형마트", "기타"];
